Added support for websocket-driver 0.6
ActionCable (from Ruby on Rails) depends on websocket-driver ~> 0.6, but
currently cuprite depends on websocket-driver ~> 0.7, which means that
curprite can't be used on rails projects using actioncable.

I've changed the version of websocket specified in the gemspec to
allow 0.6, and added an extra travis build to check that it works
correctly.