Fix deadlock on `Page.frameScheduledNavigation`
When a `Page.frameScheduledNavigation` is received the mutex is locked
but never released. According to my logs Chrome sends a
`Page.frameClearedScheduledNavigation` to notify the end of this
process.

The setup code in `Page.frameScheduledNavigation` is similar to the one
in `Page.frameStartedLoading` is received, so I applied the same cleanup
code for both ending events `Page.frameStoppedLoading` and
`Page.frameClearedScheduledNavigation`.